move services to its own subdirectory
