+++ yacs-admin-angular/src/app/department/department-detail/department-detail.component.spec.ts
@@ -2,8 +2,8 @@ import { async, tick, ComponentFixture, TestBed } from '@angular/core/testing';
 import {FormsModule} from '@angular/forms';
 import { DepartmentDetailComponent } from './department-detail.component';
 import {Department} from '../department';
-import {FakeYacsService} from '../../fake-yacs.service';
-import {InMemoryDataService} from '../../in-memory-data.service';
+import {FakeYacsService} from '../../services/fake-yacs.service';
+import {InMemoryDataService} from '../../services/in-memory-data.service';
 import {HttpClientInMemoryWebApiModule} from 'angular-in-memory-web-api';
 import {HttpClientModule} from '@angular/common/http';
 import {RouterTestingModule} from '@angular/router/testing';
