+++ yacs-admin-angular/src/app/department/department-list/department-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Department } from '../department';
-import {FakeYacsService} from '../../fake-yacs.service';
+import {FakeYacsService} from '../../services/fake-yacs.service';
 import {School} from '../../school/school';
 
 @Component({
