+++ yacs-admin-angular/src/app/school/school-detail/school-detail.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, tick, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
-import {FakeYacsService} from '../../fake-yacs.service';
-import {InMemoryDataService} from '../../in-memory-data.service';
+import {FakeYacsService} from '../../services/fake-yacs.service';
+import {InMemoryDataService} from '../../services/in-memory-data.service';
 import {HttpClientInMemoryWebApiModule} from 'angular-in-memory-web-api';
 import {HttpClientModule} from '@angular/common/http';
 import {RouterTestingModule} from '@angular/router/testing';
