+++ yacs-admin-angular/src/app/school/school-list/school-list.component.spec.ts
@@ -1,9 +1,9 @@
 import { async, tick, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SchoolListComponent } from './school-list.component';
-import {FakeYacsService} from '../../fake-yacs.service';
+import {FakeYacsService} from '../../services/fake-yacs.service';
 import {HttpClientModule} from '@angular/common/http';
-import {InMemoryDataService} from '../../in-memory-data.service';
+import {InMemoryDataService} from '../../services/in-memory-data.service';
 import {HttpClientInMemoryWebApiModule} from 'angular-in-memory-web-api';
 import {AppRouterModule} from '../../app-router/app-router.module';
 import {RouterTestingModule} from '@angular/router/testing';
