+++ yacs-admin-angular/src/app/services/in-memory-data.service.ts
@@ -1,5 +1,6 @@
 import { InMemoryDbService } from 'angular-in-memory-web-api';
-
+import {Observable } from 'rxjs/Observable';
+import{of} from 'rxjs/observable/of';
 export class InMemoryDataService implements InMemoryDbService {
 
   createDb(){
@@ -10,22 +11,22 @@ export class InMemoryDataService implements InMemoryDbService {
     const courses = [
       {id: 1, name: 'Stars, Galaxies and the Cosmos', num: '1960', department_code: 'ASTR',
         department_id: 1, min_credits: 4, max_credits: 4,
-        description: 'Check back soon'},
+        description: 'Check back soon', sections: []},
       {id: 2, name: 'Special Projects in Astronomy', num: '2940', department_code: 'ASTR',
         department_id: 1, min_credits: 1, max_credits: 6,
         description:`Study and research in various fields of astronomy to demonstrate \
         interest in and ability for independent work.Prerequisites/Corequisites: Prerequisite:\
-        permission of instructor.When Offered: Fall and spring terms annually.Credit Hours: 3`},
+        permission of instructor.When Offered: Fall and spring terms annually.Credit Hours: 3`, sections:[]},
       {id: 3, name: 'Particle Astrophysics', num: '2961', department_code: 'ASTR',
         department_id: 1, min_credits: 4, max_credits: 4,
-        description: 'Description not available...'},
+        description: 'Description not available...', sections: []},
       {id: 4, name: 'Chemistry I', num: '1100', department_code: 'CHEM',
         department_id: 4, min_credits:4, max_credits:4,
         description: `Principles of chemistry, with particular focus on atomic and molecular structure \
           and bonding, periodicity, basic thermodynamic principles, introduction to acid-base \
           chemistry and elementary chemical equilibrium, and introduction to organic chemistry. \
           Students cannot get credit for both this course and CHEM 1110. \
-          When Offered: Fall term annually.Credit Hours: 4`}
+          When Offered: Fall term annually.Credit Hours: 4`, sections: []}
     ];
     const departments=[
       {id: 1, code: 'ASTR', name: 'Astronomy', school_id: 3},
